do dns lookups in avdb scan
Since the dns lookups can take time, and we want to redo them before
every version scan in case the records have been changed, do the dns
look ups as part of the scan, instead of just once when the cell names
are added.  Do the dns lookups in parallel using multiple processes,
since the timeouts can take a while.

Reinstate the avdb add subcommand to take just a cellname an optional
description, and revert avdb import to be just for importing cellnames
and hosts from CellServDB files.